increase type check for taxConfig
